src/test/java/com/leetcode/algorithm/medium/spiralmatrixiii/SolutionTest.java
package com.leetcode.algorithm.medium.spiralmatrixiii;

import org.junit.jupiter.api.Test;

import static org.junit.jupiter.api.Assertions.assertArrayEquals;

public class SolutionTest {
  @Test
  public void testCase1() {
    Solution solution = new Solution();
    assertArrayEquals(
        new int[][] {
          {0, 0},
          {0, 1},
          {0, 2},
          {0, 3}
        },
        solution.spiralMatrixIII(1, 4, 0, 0));
  }

  @Test
  public void testCase2() {
    Solution solution = new Solution();
    assertArrayEquals(
        new int[][] {
          {1, 4}, {1, 5}, {2, 5}, {2, 4}, {2, 3}, {1, 3}, {0, 3}, {0, 4}, {0, 5}, {3, 5}, {3, 4},
          {3, 3}, {3, 2}, {2, 2}, {1, 2}, {0, 2}, {4, 5}, {4, 4}, {4, 3}, {4, 2}, {4, 1}, {3, 1},
          {2, 1}, {1, 1}, {0, 1}, {4, 0}, {3, 0}, {2, 0}, {1, 0}, {0, 0}
        },
        solution.spiralMatrixIII(5, 6, 1, 4));
  }
}
